moved encryptors to ofs/ofsencryptors module
