+++ ofsencryptors/ofsaesencryptor/functions.go
