+++ ofsencryptors/ofsaesencryptor/privatefuncs.go
@@ -3,7 +3,7 @@ package ofsaesencryptor
 import (
 	"fmt"
 
-	ofsmessages "github.com/thomas-osgood/ofs/server/internal/messages"
+	encryptormessages "github.com/thomas-osgood/ofs/ofsencryptors/internal/messages"
 )
 
 // function designed to validate an AES encryptor's key.
@@ -11,7 +11,7 @@ func validateKey(key []byte) (err error) {
 	var keylen int = len(key)
 
 	if (keylen != 32) && (keylen != 24) && (keylen != 16) {
-		return fmt.Errorf(ofsmessages.ERR_KEYLEN_AES)
+		return fmt.Errorf(encryptormessages.ERR_AES_KEYLEN)
 	}
 
 	return nil
